refactor: JpaRepository 인터페이스 패키지 분리
