+++ src/main/java/roomescape/reservation/repository/jpa/ThemeJpaRepository.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.repository;
+package roomescape.reservation.repository.jpa;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -7,6 +7,7 @@ import org.springframework.data.repository.query.Param;
 import roomescape.reservation.domain.Theme;
 import roomescape.reservation.domain.ThemeName;
 import org.springframework.data.jpa.repository.JpaRepository;
+import roomescape.reservation.repository.ThemeRepository;
 
 public interface ThemeJpaRepository extends JpaRepository<Theme, Long>, ThemeRepository {
